Setup ability to handle games
