+++ app.js
@@ -119,6 +119,7 @@ global.Punishments = require('./punishments');
 global.Chat = require('./chat');
 global.Rooms = require('./rooms');
 
+global.Console = require('./console.js');
 global.Tells = require('./tells.js');
 
 delete process.send; // in case we're a child process
